+++ src/plugins/cancel.js
@@ -1,14 +1,16 @@
 import axios from 'axios'
 let cancel
+const urlhistory = new Set()
 // 防止多次点击发起重复的请求
 axios.interceptors.request.use(
   config => {
-    if (typeof cancel === 'function') {
+    if (typeof cancel === 'function' && urlhistory.has(config.url)) {
       cancel('取消请求')
     }
     config.cancelToken = new axios.CancelToken((c) => {
       cancel = c
     })
+    urlhistory.add(config.url)
     return config
   })
 axios.interceptors.response.use(
